Implemented disposition for streams.
Streams now can take a disposition which can be used to distinguish inline streams and attachment streams. The idea here is that inline streams are used for internal functionality in an application, and attachment streams are streams requested by the user. This can be used with progress reporting to decide on what streams to report progress.

Breaking change: the web stream API doesn't take a disposition query string argument anymore. This has been replaced with this new functionality.